pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "cbs-eval"
version = "0.1.0"
description = "CBS dynamic scheduling evaluation framework for LLM inference"
requires-python = ">=3.10"
dependencies = [
    "simpy>=4.0",
    "numpy>=1.24",
    "scipy>=1.10",
    "matplotlib>=3.7",
    "pyyaml>=6.0",
    "aiohttp>=3.8",
    "transformers>=4.36",
]

[project.optional-dependencies]
gpu = ["vllm>=0.18.0"]

[tool.setuptools.packages.find]
include = ["mlwd*", "real*", "sim*", "workload*", "analysis*", "plot*"]
